Don't depend on the zip name for the image name
Don't expect the image to have the same name as the archive. There are
cases where those don't match. If you use an outdated URL,  bitnami will
fetch you with an archive that contains a newer image that the one
you've asked for, without redirecting. In this case the content of the
archive will have a different version in the file names than the one
the archive has.